pyproject.toml
[project]
name = "iam-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.124.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.45",
    "aiosmtplib>=5.0.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=8.2",
    "cryptography>=46.0.3",
    "aioredis>=2.0.1",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
anyio_backend = "asyncio"
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "pytest-mock>=3.15.1",
    "mypy>=1.19.1",
    "coverage>=7.13.4",
    "ty>=0.0.29",
]

[tool.ruff]
line-length = 120
indent-width = 4

src = ["app"]
 
[tool.ruff.lint]
preview = true

select = [
    "UP",   # pyupgrade
    "I",    # isort
    "E",    # flake8-errors
    "W",    # flake8-warnings
    "F",    # flake8-pyflakes
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "T20",  # flake8-print
    "T10",  # flake8-debugger
    "PT",   # flake8-pytest-style
    "RUF",  # Ruff
    "DTZ",  # flake8-datetimez
    "ERA",  # eradicate
    "FURB", # refurb
]

ignore = [
    "C408",  # Ненужный вызов dict, list и т.д.
    "RUF001", # Игнорирование похожих символов с разными ASCII
    "RUF002", # Игнорирование похожих символов с разными ASCII
    "RUF003", # Игнорирование похожих символов с разными ASCII
    "S110",  # try-except-pass
]

exclude = [
    "tests/**",
    "alembic/**"
]
fixable = [
    "I", # sorting imports
    "E", # flake8-errors
    "F401",
    "RUF010",
    "UP035",
    "UP043",
    "UP017",
    "E501",
    "B008"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.flake8-bandit]
check-typed-exception = true

[tool.ruff.lint.isort]
known-first-party = ["app"]
